installerplanner/api/requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
openpyxl
xlsxwriter
python-dateutil
requests

